requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-Caching==2.0.2
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
python-dateutil==2.8.2